fastapi_services/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
sentence-transformers==2.7.0
numpy==1.26.4
chromadb==0.5.0
pinecone-client==4.1.0
torch==2.3.0
